requirements.txt
streamlit
pandas
numpy
sqlalchemy
psycopg2-binary
loguru
plotly
dotenv
bcrypt
matplotlib
matplotx
anywidget